+++ test/utils/gameWrapper.ts
@@ -14,7 +14,6 @@ import MockImage from "#test/utils/mocks/mocksContainer/mockImage";
 import MockTextureManager from "#test/utils/mocks/mockTextureManager";
 import fs from "fs";
 import Phaser from "phaser";
-import InputText from "phaser3-rex-plugins/plugins/inputtext";
 import { vi } from "vitest";
 import { MockGameObjectCreator } from "./mocks/mockGameObjectCreator";
 import InputManager = Phaser.Input.InputManager;
@@ -26,54 +25,7 @@ import UpdateList = Phaser.GameObjects.UpdateList;
 import { version } from "../../package.json";
 import { MockTimedEventManager } from "./mocks/mockTimedEventManager";
 
-Object.defineProperty(window, "localStorage", {
-  value: mockLocalStorage(),
-});
-Object.defineProperty(window, "console", {
-  value: mockConsoleLog(false),
-});
-
-
-InputText.prototype.setElement = () => null;
-InputText.prototype.resize = () => null;
-Phaser.GameObjects.Image = MockImage;
-window.URL.createObjectURL = (blob: Blob) => {
-  blobToString(blob).then((data: string) => {
-    localStorage.setItem("toExport", data);
-  });
-  return null;
-};
-navigator.getGamepads = () => [];
-global.fetch = vi.fn(MockFetch);
-Utils.setCookie(Utils.sessionIdKey, 'fake_token');
-
-
-window.matchMedia = () => ({
-  matches: false,
-});
-
-
-/**
- * Sets this object's position relative to another object with a given offset
- * @param guideObject {@linkcode Phaser.GameObjects.GameObject} to base the position off of
- * @param x The relative x position
- * @param y The relative y position
- */
-const setPositionRelative = function (guideObject: any, x: number, y: number) {
-  const offsetX = guideObject.width * (-0.5 + (0.5 - guideObject.originX));
-  const offsetY = guideObject.height * (-0.5 + (0.5 - guideObject.originY));
-  this.setPosition(guideObject.x + offsetX + x, guideObject.y + offsetY + y);
-};
-
-Phaser.GameObjects.Container.prototype.setPositionRelative = setPositionRelative;
-Phaser.GameObjects.Sprite.prototype.setPositionRelative = setPositionRelative;
-Phaser.GameObjects.Image.prototype.setPositionRelative = setPositionRelative;
-Phaser.GameObjects.NineSlice.prototype.setPositionRelative = setPositionRelative;
-Phaser.GameObjects.Text.prototype.setPositionRelative = setPositionRelative;
-Phaser.GameObjects.Rectangle.prototype.setPositionRelative = setPositionRelative;
-
-
-export default class GameWrapper {
+export class GameWrapper {
   public game: Phaser.Game;
   public scene: BattleScene;
